+++ CHANGELOG.md
@@ -138,7 +138,7 @@
 
 ## 0.7.0 (Apr 30, 2017)
 
-- Changed the `main` field value from `dist/viewer.js` (UMD) to `dist/viewer.common.js` (CommonJS).
+- Changed the `main` field value from `dist/viewerjs.js` (UMD) to `dist/viewerjs.common.js` (CommonJS).
 - Added `module` and `browser` fields to `package.json`.
 - Fixed an issue of touch zoom.